Decouple MythTV configure from FFmpeg configure, first step
Instead of having MythTV configure do all the work that is needed for
FFmpeg, rather have MythTV configure invoke the FFmpeg configure.
This eliminates the need to keep updating MythTV configure with FFmpeg
changes.

This is step 1. FFmpeg configure is invoked and MythTV configure does
none of the FFmpeg work. There is still much unneeded FFmpeg
code in the MythTV configure.

Refs #13259